Move validate privateKey & seedPhrase functionality to dedicated functions
