+++ ethdk/test/ExternallyOwnedAccount/ExternallyOwnedAccount.test.ts
@@ -175,140 +175,4 @@ describe('ExternallyOwnedAccount', () => {
       expect(result).to.equal('1.23')
     })
   })
-
-  describe('isPrivateKey', () => {
-    it('should return true for a valid private key', () => {
-      // Arrange
-      const privateKey =
-        '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80'
-
-      // Act
-      const result = ExternallyOwnedAccount.validatePrivateKey(privateKey)
-
-      // Assert
-      expect(result.success).to.equal(true)
-      expect(result.error).to.equal('')
-    })
-
-    const testCases = [
-      {
-        privateKey:
-          '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff800',
-        errorMessage:
-          'Private key must be 64 characters long, excluding the "0x" prefix',
-      },
-      {
-        privateKey:
-          '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff8',
-        errorMessage:
-          'Private key must be 64 characters long, excluding the "0x" prefix',
-      },
-      {
-        privateKey:
-          'ac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80',
-        errorMessage: 'Private key must start with "0x"',
-      },
-      {
-        privateKey: '0x',
-        errorMessage:
-          'Private key must be 64 characters long, excluding the "0x" prefix',
-      },
-      {
-        privateKey: '0x00',
-        errorMessage:
-          'Private key must be 64 characters long, excluding the "0x" prefix',
-      },
-      {
-        privateKey: 'not a private key',
-        errorMessage: 'Private key must start with "0x"',
-      },
-      {
-        privateKey:
-          '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2fxyz',
-        errorMessage:
-          'Private key must only contain lowercase or uppercase hexadecimal characters (a-f, A-F, 0-9)',
-      },
-      {
-        privateKey:
-          '0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2f!&#',
-        errorMessage:
-          'Private key must only contain lowercase or uppercase hexadecimal characters (a-f, A-F, 0-9)',
-      },
-    ]
-    testCases.forEach((testCase) => {
-      it(`should return false for an invalid private key: ${testCase.privateKey}`, () => {
-        // Arrange & Act
-        const result = ExternallyOwnedAccount.validatePrivateKey(
-          testCase.privateKey,
-        )
-        // Assert
-        expect(result.success).to.equal(false)
-        expect(result.error).to.equal(testCase.errorMessage)
-      })
-    })
-  })
-
-  describe('isSeedPhrase', () => {
-    const validTestCases = [
-      'sock poet alone around radar forum quiz session observe rebel another choice',
-      'test test test test test test test test test test test junk',
-    ]
-    validTestCases.forEach((testCase) => {
-      it(`should return true for a valid seed phrase: ${testCase}`, () => {
-        // Arrange & Act
-        const result = ExternallyOwnedAccount.validateSeedPhrase(testCase)
-
-        // Assert
-        expect(result.success).to.equal(true)
-        expect(result.error).to.equal('')
-      })
-    })
-
-    const invalidTestCases = [
-      {
-        seedPhrase:
-          'sock poet alone around radar forum quiz session observe rebel another choice long',
-        error: 'Seed phrase must contain exactly 12 words',
-      },
-      {
-        seedPhrase:
-          'sock poet alone around radar forum quiz session observe rebel short',
-        error: 'Seed phrase must contain exactly 12 words',
-      },
-      {
-        seedPhrase:
-          'sock poet alone around radar forum quiz session observe rebel another numb3r',
-        error:
-          'Each word in the seed phrase must only contain lowercase letters (a-z)',
-      },
-      {
-        seedPhrase:
-          'sock poet alone around radar forum quiz session observe rebel another whitespa ce',
-        error: 'Seed phrase must contain exactly 12 words',
-      },
-      {
-        seedPhrase:
-          'Sock Poet Alone Around Radar Forum Quiz Session Observe Rebel Another Captials',
-        error:
-          'Each word in the seed phrase must only contain lowercase letters (a-z)',
-      },
-      {
-        seedPhrase:
-          'sock poet alone around radar forum quiz session observe rebel another  extraspace',
-        error: 'Seed phrase must contain exactly 12 words',
-      },
-    ]
-    invalidTestCases.forEach((testCase) => {
-      it(`should return true for an invalid seed phrase: ${testCase.seedPhrase}`, () => {
-        // Arrange & Act
-        const result = ExternallyOwnedAccount.validateSeedPhrase(
-          testCase.seedPhrase,
-        )
-
-        // Assert
-        expect(result.success).to.equal(false)
-        expect(result.error).to.equal(testCase.error)
-      })
-    })
-  })
 })
